Convert to int64 before conversion to string

+++ pkg/postgresql/postgresql.go
@@ -171,7 +171,7 @@ func (p *Manager) Start() error {
 	}
 	name := filepath.Join(p.pgBinPath, "pg_ctl")
 	cmd := exec.Command(name, "start", "-w", "-t",
-		string(startTimeout/time.Second), "-D", p.dataDir)
+		string(int64(startTimeout/time.Second)), "-D", p.dataDir)
 	stdoutPipe, err := cmd.StdoutPipe()
 	if err != nil {
 		return trace.Wrap(err)
